feat: Refactored utils, moved function to calculate file size (#170)
* feat: Refactored utils, moved function to calculate file size

* Update src/unstract/sdk/utils/common_utils.py


---------
